coMMit before entering a new database

+++ data/src/main/java/com/BySandS/testsandquizes/data/test/storage/QuestionStorage.kt
@@ -3,5 +3,5 @@ package com.BySandS.testsandquizes.data.test.storage
 import com.BySandS.testsandquizes.data.test.storage.models.QuestionModelDb
 
 interface QuestionStorage {
-    fun getQuestionsList(difficultyId: Long, quantityOfQuestions: Int): List<QuestionModelDb>
+    fun getQuestionsList(difficultyId: Long, quantityOfQuestions: Long): List<QuestionModelDb>
 }
